membuat kondisional pada URL PUT

+++ src/main/java/com/askar/pws/askar101/ProductServiceController.java
@@ -47,10 +47,16 @@ public class ProductServiceController {
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id")String id, @RequestBody Product product){
-       productRepo.remove(id);
-       product.setId(id);
-       productRepo.put(id, product);
-       return new ResponseEntity<>("Product is updated Successfully", HttpStatus.OK);
+       if(!productRepo.containsKey(id)){
+           return new ResponseEntity<>("Product is not assigned", HttpStatus.OK);
+       }else{
+           productRepo.remove(id);
+           product.setId(id);
+           productRepo.put(id, product);
+           return new ResponseEntity<>("Product is updated Successfully", HttpStatus.OK);
+           
+       }
+       
    }
    @RequestMapping(value = "/products", method = RequestMethod.POST)
    public ResponseEntity<Object> createProduct(@RequestBody Product product) {
